rename: 파일 이름 수정
Http 관련 enum을 http.constant 로 이동
requestParser를 webserver로 이동
