+++ src/main/java/http/HttpRequestHandler.java
@@ -1,5 +1,7 @@
 package http;
 
+import http.constant.HttpMethod;
+import http.constant.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import util.FileUtils;
