+++ src/main/java/http/HttpResponse.java
@@ -1,5 +1,7 @@
 package http;
 
+import http.constant.HttpHeader;
+import http.constant.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import util.MimeType;
