+++ src/main/java/http/constant/HttpMethod.java
@@ -1,4 +1,4 @@
-package http;
+package http.constant;
 
 public enum HttpMethod {
     GET,
